refactor: 過度な設定抽象化を削除して簡潔性を重視
## 🎯 保守性重視の簡素化

### ❌ 削除した不必要な拡張性
- 複雑なTwitter設定オプション（theme, dnt, conversation等）
- デバッグフラグとタイムアウト設定
- 使用しないTwitter API型定義（createTweet等）
- 2つしかないパッケージの共通設定ファイル

### ✅ 簡潔になった実装
- Twitter設定: 1つのbooleanフラグのみ
- 型定義: サービスファイル内に統合（10行）
- タイムアウト: 15秒固定
- パッケージ設定: 直接記述

## 📊 改善効果

 < /dev/null |  項目 | Before | After | 改善 |
|------|--------|-------|------|
| 設定オプション | 7個 | 1個 | **85%削減** |
| Twitter型定義 | 100行 | 10行 | **90%削減** |
| 設定ファイル | 3個 | 1個 | **67%削減** |
| 保守箇所 | 分散 | 集約 | **一元化** |

## 原則
- YAGNI: 実際に使わない機能は実装しない
- 固定値: 変更しない設定は定数化
- 簡潔性: 過度な抽象化より分かりやすさ
